+++ tests/calculators/test_hold_calculator.py
@@ -1,5 +1,5 @@
 from decimal import Decimal
-from sportsbooklib.calculators.get_hold import get_hold_for_selections
+from sportsbooklib.calculators.hold_calc import get_hold_for_selections
 from sportsbooklib.models.odds.enums import OddsFormat
 from sportsbooklib.models.odds.odds import Odds
 from sportsbooklib.models.selection.selection import Selection
